Restructure utility folders and move files to common file, #11
