+++ components/common/hooks/useFetcher.ts
@@ -1,5 +1,5 @@
 import { useCallback } from "react";
-import { fetchData } from "../../pages/api/fetchData";
+import { fetchData } from "../../../pages/api/fetchData";
 
 export const useFetcher = () => {
     const data = useCallback(fetchData, [])
