+++ components/common/hooks/useImagePlaceholder.ts
