+++ components/common/types/catImageResponse.types.ts
