+++ components/common/types/catResponse.types.ts
