+++ components/common/utils/localStorageUtils.ts
@@ -1,4 +1,4 @@
-import { CardInterface } from "../Card/card.types";
+import { CardInterface } from "../../Card/card.types";
 
 export const handlerFavorite = (
   cardData: CardInterface,
