+++ components/common/utils/propertyUtils.ts
