+++ pages/favorites/index.tsx
@@ -1,8 +1,8 @@
 import { CardList } from "@/components/Card/CardList";
 import { CardInterface } from "@/components/Card/card.types";
-import { withDataCheck } from "@/components/hocs/withDataCheck";
-import { useFavoriteStorage } from "@/components/hooks/useFavoriteStorage";
-import { removeFavoriteItem } from "@/components/utils/localStorageUtils";
+import { withDataCheck } from "@/components/common/hocs/withDataCheck";
+import { useFavoriteStorage } from "@/components/common/hooks/useFavoriteStorage";
+import { removeFavoriteItem } from "@/components/common/utils/localStorageUtils";
 const EnhancedCardList = withDataCheck(CardList);
 
 const Favorites = () => {
